api/requirements.txt
asgiref==3.4.0
Django==3.2.23
django-extensions
django-cors-headers==3.10.0
django-environ==0.8.1
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1